Lite: buildSrc for version, plugin and dependency management
Refactoring the build gradle file to now use the dependency, plugin and version information located in the buildSrc module. This allows us to now have a central point for upgrading versions and dependencies.

In addition to that, adding dependencies for mockk and assertK, which will be used to write more light weighted unit tests and have a better dsl for writing assertions.

Also adding the arrow stack, especially arrow core to the project, as this will be used for error handling. We are going down the functional programming road when it comes to error handling, as we have a better control flow and readability this way, as we are not throwing exceptions, but working with the Either result type. This way we just need to follow the code path in order to know what's happening and not need to hunt down exceptions handlers located somewhere, if at all, in the project.