Lite[Invoice|Customer]: Data classes as entities
Refactoring the entity classes to kotlin data classes and making the properties immutable. We should favor immutability over mutability in order
to avoid side effects. Mutability should be something where developers should put in some thought before implementing it, as it could also
lead to inconsistent or wrong data, for example when reassigning and id to a new value.

The values are all still nullable, which is not clean yet, and will be refactored down the line. Nullability is part of the business logic, and not just a plain
refactoring, and needs some careful thoughts.