Lite: Basic structural refactoring
Refactoring the given code structure to match more to a hexagonal approach.
First of all the customer domain type is now located in a separate domain package, as it reflects a separate domain. The structure inside
a domain package is now the following:

On root level of a domain package, the domain type and use cases are defined. Making this information / code top level of a domain package, gives
a developer an idea of the capabilities of the package, without the need of diving into the whole structure of the domain package. In order to
keep it simple, there is a service and adapter package. The service package is meant for the service implementation (middle ware), where the adapter
package is meant for persistence, rest and so on adapters.

Business Logic goes into the service package, and functions / methods are always defined in one or many use case interfaces.
Database / persistence code, as well as rest controllers or other code that connects to the infrastructure layer, goes into the adapter package.

There could be done some additional fine-tuning, like separating write / read use cases and also cleaning up the pure domain logic, freeing it from any
kind of framework / library code, to make it portable and independent of infrastructure code. But for this task I would consider it an overkill.