Lite[Invoice|Failure]: Introducing arrow-kt for error handling
Using arrow-kt, especially the either type as an alternative to traditional try catch error handling. The benefit of using either in this case, is that
we are forced to handle possible issues linear in our code. With classic exceptions, one can define spring exception handlers in different places
or also may don't catch an exception using try catch at all. Especially as kotlin doesn't use typed exceptions.

Using either, we are 'trapping' on  the lowest level possible, which means, spanning a context where a side effect can occur (Spring Data Repositories). Trapping is nothing else, than spanning a block that takes care of transforming an exception thrown by spring into an either.

The either type is evaluated on the top level (view -> rest api). This also gives us fine granted control of the status code and information we want to use as
response.

Right now the project is setup using classic spring boot starter web, which doesn't support async coding well. ArrowKT on the other side is using kotlin
suspend functions to model monads. This is a bit of a conflict here, but we'll solved it for now, by wrapping the call of a suspend function using a runBlocking block in the view / api layer.