Lite: Authentication and Authorization for the backend
Implementing basic auth for the backend service. Using an in memory user details manager to keep it simple for now. There are 
two users defined: jane and john. Alongside every user has a role: USER (CRUD), VIEWER (Read-Only).

In order to request an api endpoint, one must submit basic auth credentials, and must make sure to have the appropriate role to perform the action.

In the future, one should switch the in memory implementation for something like keycloak or auth0.