Lite[Invoice]: Not all fields should be settable from the outside
Implementing a view model on the adapter layer to limit the fields one can set from the outside for an invoice. A client should never be able to
set attributes like id, status, creationDate, priceGross (could be calculated), and so on.

In order to accomplish this, a view model is used to parse the request body. The view model is translated using an extension function into the actual
domain object. 

Also redefining nullability for the invoice type. Before, all fields where nullable and had null as a default value. This is now changed, where 
fields are non nullable if they're required, and also providing meaningful defaults for fields values, like setting an initial status value always to open.
For the invoice status, an enum is now in place to avoid issues with typos in the status field, as this field will be required down the line to filter for
certain invoices having a certain state.