Lite: Implementing basic auditing for customer and invoice
Activating / Enabling and implementing a basic jpa auditing. Now for update / insert, the logged in username and the current date-time is injected by jpa auditing into the invoice or customer model and stored to the database. Right now an interface `Auditable` is in place, specifying the fields required for auditing. This is the common reusable part, implemented by every jpa entity, that should be audited.